error handling is almost there, needs proper
