+++ includes/ft_42sh.h
@@ -6,7 +6,7 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: Invalid date        by                   #+#    #+#             */
-/*   Updated: 2023/02/11 12:12:52 by mviinika         ###   ########.fr       */
+/*   Updated: 2023/02/11 15:49:08 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -621,7 +621,7 @@ char			*remove_braces(char *str);
 char			*get_value(t_shell *sh, char *var, char *subst, int format);
 int				format_mode(char op);
 int				join_values(t_shell *sh, t_param *pa, char *cmd, int ret);
-char			*get_operator(char *cmd, int *ret);
+char			*get_operator(char *cmd);
 int				is_param_exp_char(char *flag);
 int				splitter(char *cmd, t_param *pa, int *ret);
 int				expander(t_param *pa, int ret);
@@ -639,6 +639,7 @@ char			*find_from_end(char *haystack, char *needle);
 char			*find_from_begin_glob(char *haystack, char *needle);
 int				is_substring_id(char *needle);
 int				check_var_validity(char *var);
+int 			check_substitutions(char *cmd, int *ret, t_param *pa);
 
 /*			  		 SIGNALS				*/
 void			signal_exec(int num);
