+++ sources/parameter_formatting/is_param_exp_char.c
@@ -6,7 +6,7 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/10 11:15:36 by mviinika          #+#    #+#             */
-/*   Updated: 2023/02/11 11:20:51 by mviinika         ###   ########.fr       */
+/*   Updated: 2023/02/11 16:51:02 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@ int is_param_exp_char(char *flag)
 		return (0);
 	else if (flag[0] == ':')
 		return (flag[1] == '-' || flag[1] == '+'
-			|| flag[1] == '=' || flag[1] == '?');
+			|| flag[1] == '=');
 	return ((flag[0] == '%' && flag[1]) || (flag[0] == '#' && flag[1]));
 }
